Add files via upload
Stemming and Lemmatization using NLP.
In the field of Natural Language Processing (NLP) Lemmatization and Stemming are Text Normalization techniques. 
These techniques are used to prepare words, text, and documents for further processing.
Languages such as English, Hindi consists of several words which are often derived from one another. 
Further, Inflected Language is a term used for a language that contains derived words.